requirements.txt
# Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Templates
jinja2>=3.1.3

# Supabase
supabase>=2.3.4

# Validación y configuración
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.1.0

# Autenticación
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Generación de PDFs
reportlab>=4.0.9

# Envío de correos
resend>=0.7.0

# Tareas programadas
apscheduler>=3.10.4

# Utilidades
python-dateutil>=2.8.2
requests>=2.31.0
